Add fallback Tor friendly API for Bitcoin fee rates
